pyproject.toml
[project]
name = "price-comparison-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "crawl4ai>=0.6.3",
    "fastapi>=0.116.0",
    "google-search-results>=2.4.2",
    "httpx>=0.28.1",
    "lxml>=5.4.0",
    "playwright>=1.53.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
